AndroidBOINC: Fix multiplication of data size factors (daily transfer & reserved free storage space)

+++ android/BOINC/src/edu/berkeley/boinc/adapter/PrefsListAdapter.java
@@ -101,10 +101,10 @@ public class PrefsListAdapter extends ArrayAdapter<PrefsListItemWrapper>{
 	    					value = NumberFormat.getInstance().format(item.status) + this.activity.getString(R.string.prefs_unit_celsius);
 	    					break;
 	    				case MEGABYTES:
-	    					value = Formatter.formatShortFileSize(this.activity, item.status.longValue() * 0x100000);
+	    					value = Formatter.formatShortFileSize(this.activity, (long)(item.status.doubleValue() * 0x100000));
 	    					break;
 	    				case GIGABYTES:
-	    					value = Formatter.formatShortFileSize(this.activity, item.status.longValue() * 0x40000000);
+	    					value = Formatter.formatShortFileSize(this.activity, (long)(item.status.doubleValue() * 0x40000000));
 	    					break;
 	    				default:
 	    					value = NumberFormat.getInstance().format(item.status);
